values and vregs for which we have no register class information don't default to ints now (as this was causing empty register sets when trying to coalesce an int and a double). It seems to work now, although painfully slowly.

+++ mach/proto/mcg/treebuilder.c
@@ -1859,7 +1859,7 @@ static void emit_jumptable(struct ir* targetvalue, struct jumptable* jumptable)
 	materialise_stack();
     while (hashtable_next(&jumptable->targets, &hit))
 	{
-		int value = (int) hit.key;
+		int value = (int)(intptr_t) hit.key;
 		struct basicblock* target = hit.value;
 		struct basicblock* nextblock = bb_get(NULL);
